move extension points to node package
