+++ plugins/backend/rag-ai-node/src/index.ts
@@ -14,3 +14,4 @@
  * limitations under the License.
  */
 export * from './types';
+export * from './extensions';
